init v0.4.8

+++ crates/parser/src/common/concept.rs
@@ -5,7 +5,7 @@
 
 use std::hash::Hash;
 
-use serde::Serialize;
+use serde::{Deserialize, Serialize};
 
 use crate::{
     common::{connection::Connection, enum_def::Protocol, util::date_str, FastHashMap, Instance, NString},
@@ -23,6 +23,7 @@ pub type ConversationKey = (u64, u64);
 
 pub type Timestamp = u64;
 
+#[derive(Deserialize, Serialize)]
 pub struct Criteria {
     pub size: usize,
     pub start: usize,
@@ -111,7 +112,7 @@ impl CounterItem {
     }
 }
 
-#[derive(Serialize)]
+#[derive(Serialize, Debug)]
 pub struct ProgressStatus {
     pub total: usize,
     pub cursor: usize,
@@ -201,7 +202,7 @@ impl HttpHeadContinue {
     }
 }
 
-#[derive(Default, Clone, Serialize)]
+#[derive(Default, Clone, Serialize, Deserialize)]
 pub struct Field {
     // #[serde(skip)]
     // pub extra_data: Option<Vec<u8>>,
@@ -517,7 +518,7 @@ impl DNSResponse {
         rs.latency = period(rs.res_ts, rs._latency);
 
         if let Some(index) = item.request {
-            if let Some(frame) = instance.frame(index as usize) {
+            if let Some(frame) = instance.frame(index) {
                 // start = frame.info.time;
                 // rs.ts = start;
                 // rs.ts_str = date_str(rs.res_ts);
@@ -530,7 +531,7 @@ impl DNSResponse {
         }
 
         if let Some(index) = item.response {
-            if let Some(frame) = instance.frame(index as usize) {
+            if let Some(frame) = instance.frame(index) {
                 if let Some((ip, _)) = frame.addresses(instance.context()) {
                     rs.target = ip;
                 }
@@ -598,7 +599,7 @@ pub enum HttpEncoding {
     Zstd,
 }
 
-#[derive(Debug)]
+#[derive(Serialize, Debug)]
 pub struct HttpMessageDetail {
     pub is_request: bool,
     pub headers: Vec<String>,
